+++ CHANGELOG.md
@@ -30,11 +30,12 @@ Feature:
 Fix:
 - Track ID of delta lineage started with 0, leading to a lost cell
 - Fixed double assignment bug in delta tracking, where the same cell can be overwritten, i.e. be the successor for two cells
+- Delta tracking added connectivity to lineage
 
 Efficiency:
 - Delta tracking now avoids resize of input
 - Delta tracking data prep and cleanup much more efficient
-- Delta tracking lower memory footprint 
+- Delta tracking lower memory footprint
 
 ## [0.3.4]